requirements.txt
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
webdriver-manager==4.0.1
wsproto==1.2.0
